requirements.txt
anthropic>=0.40.0
playwright>=1.48.0
python-dotenv>=1.0.0
flask>=3.0.0
